pyproject.toml
[tool.poetry]
name = "adb-setup"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
notebook = "^6.4.12"
requests = "^2.28.1"
urlpath = "^1.2.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^22.6.0", allow-prereleases = true}
pydocstyle = "^6.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
